requirements.txt
fastapi-users >= 9.1.0
tortoise-orm >= 0.17.6,<0.18.0
